meethub/requirements.txt
Django>=1.10.5,<1.11
django-bootstrap3-datetimepicker==2.2.3
django-disqus==0.5
django-friendship==1.5.0
django-map-widgets==0.1.8
django-registration==2.2
Pillow>=4.1.0
-e git+https://github.com/GabrielUlici/django-bootstrap4#egg=django-bootstrap4
psycopg2==2.7.1
pygeocoder==1.2.5
